requirements.txt
streamlit>=1.38
pandas>=2.1
numpy>=1.26
plotly>=5.22
matplotlib>=3.8
fpdf2>=2.7
openai>=1.34
pydantic>=2.7
pytest>=8.3

